Reorganice code acording to ES6 best practices
